pyproject.toml
[build-system]
requires = ["setuptools>=65.5.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-kitchen"
version = "0.1.0"
description = "FullStack CRUD app written in Django & DRF."
authors = [
    {name = "akiko23", email = "dmvasilenko43@gmail.com"},
]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "Django==5.0.4",
    "djangorestframework==3.15.1",
    "drf-spectacular==0.27.2",
    "psycopg2==2.9.9",
]

[project.optional-dependencies]
lint = [
    "mypy==1.7.1",
    "isort==5.12.0",
    "ruff==0.1.6",
]

[project.urls]
"Homepage" = "https://github.com/akiko23/django-kitchen"
"Bug Tracker" = "https://github.com/akiko23/django-kitchen/issues"


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "kitchen_app/migrations",
]

# Same as Black.
line-length = 98
indent-width = 4

# Assume Python 3.11
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "single"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"


# Isort configuration
[tool.isort]
skip = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    "build",
    "dist",
    "kitchen_app/migrations",
    "venv",
]
include_trailing_comma = true
use_parentheses = true
multi_line_output = 3


[tool.mypy]
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    "build",
    "dist",
    "kitchen_app/migrations",
    "venv",
]
strict = true
python_version = "3.11"
ignore_missing_imports = true
install_types = true
non_interactive = true
plugins = [
    "mypy_django_plugin.main",
]

# from https://blog.wolt.com/engineering/2021/09/30/professional-grade-mypy-configuration/
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
show_error_codes = true
warn_unused_ignores = false

disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_any_unimported = false

[tool.django-stubs]
django_settings_module = "myproject.settings"